# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: List daily records for current student
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyRecord"
    post:
      operationId: createRecord
      tags: [records]
      summary: Create a new daily record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a single daily record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /records/extract-text:
    post:
      operationId: extractFromText
      tags: [records]
      summary: Extract anxiety/sleep/scores from free text via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractTextInput"
      responses:
        "200":
          description: Extracted values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedValues"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /exam/generate:
    post:
      operationId: generateExam
      tags: [exam]
      summary: Generate adaptive exam based on student weaknesses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateExamInput"
      responses:
        "200":
          description: Exam questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /exam/submit:
    post:
      operationId: submitExam
      tags: [exam]
      summary: Submit exam answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSubmitInput"
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    DailyRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        recordDate:
          type: string
        mathScore:
          type: ["number", "null"]
        physicsScore:
          type: ["number", "null"]
        chemistryScore:
          type: ["number", "null"]
        biologyScore:
          type: ["number", "null"]
        anxietyLevel:
          type: ["integer", "null"]
        sleepHours:
          type: ["number", "null"]
        focusDrop:
          type: ["boolean", "null"]
        socialInteraction:
          type: ["integer", "null"]
        examDaysLeft:
          type: ["integer", "null"]
        freeText:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, recordDate, createdAt]
    DailyRecordInput:
      type: object
      properties:
        recordDate:
          type: string
        mathScore:
          type: ["number", "null"]
        physicsScore:
          type: ["number", "null"]
        chemistryScore:
          type: ["number", "null"]
        biologyScore:
          type: ["number", "null"]
        anxietyLevel:
          type: ["integer", "null"]
        sleepHours:
          type: ["number", "null"]
        focusDrop:
          type: ["boolean", "null"]
        socialInteraction:
          type: ["integer", "null"]
        examDaysLeft:
          type: ["integer", "null"]
        freeText:
          type: ["string", "null"]
      required: [recordDate]
    ExtractTextInput:
      type: object
      properties:
        text:
          type: string
      required: [text]
    ExtractedValues:
      type: object
      properties:
        anxietyLevel:
          type: ["integer", "null"]
        sleepHours:
          type: ["number", "null"]
        mathScore:
          type: ["number", "null"]
        physicsScore:
          type: ["number", "null"]
        chemistryScore:
          type: ["number", "null"]
        biologyScore:
          type: ["number", "null"]
        examDaysLeft:
          type: ["integer", "null"]
        summary:
          type: string
      required: [summary]
...
    Question:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        topic:
          type: ["string", "null"]
        difficulty:
          type: ["integer", "null"]
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
      required: [id, subject, text, optionA, optionB, optionC, optionD, correctAnswer]
...
    GenerateExamInput:
      type: object
      properties:
        count:
          type: integer
          default: 10
      required: []
    ExamSubmitInput:
      type: object
      properties:
        sessionId:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ExamAnswer"
      required: [sessionId, answers]
    ExamAnswer:
      type: object
      properties:
        questionId:
          type: integer
        chosen:
          type: string
      required: [questionId, chosen]
    ExamResult:
      type: object
      properties:
        sessionId:
          type: string
        totalQuestions:
          type: integer
        correctCount:
          type: integer
        score:
          type: number
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectScore"
      required: [sessionId, totalQuestions, correctCount, score, subjectBreakdown]
    SubjectScore:
      type: object
      properties:
        subject:
          type: string
        total:
          type: integer
        correct:
          type: integer
      required: [subject, total, correct]
...