# lib/api-zod/src/generated/types/dailyRecordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */

export interface DailyRecordInput {
  recordDate: string;
  /** @nullable */
  mathScore?: number | null;
  /** @nullable */
  physicsScore?: number | null;
  /** @nullable */
  chemistryScore?: number | null;
  /** @nullable */
  biologyScore?: number | null;
  /** @nullable */
  anxietyLevel?: number | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  focusDrop?: boolean | null;
  /** @nullable */
  socialInteraction?: number | null;
  /** @nullable */
  examDaysLeft?: number | null;
  /** @nullable */
  freeText?: string | null;
}