# lib/api-zod/src/generated/types/subjectScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ترنم مهر - سیستم هوشمند پایش سلامت روان و تحصیلی دانش‌آموزان
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectScore {
  subject: string;
  total: number;
  correct: number;
}